updated user controller

+++ src/main/java/com/example/simplespringapp/repositories/UserRepository.java
@@ -8,7 +8,7 @@ import com.example.simplespringapp.user.Users;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users, Long>{
+public interface UserRepository extends JpaRepository<Users, Integer>{
 
     Optional<Users> findByUsername(String username);
